+++ NEWS.md
@@ -1,3 +1,15 @@
-# GDPuc 0.5.0
+# GDPuc 0.7.0
+
+* Add argument `return_cfs` to return the conversion factors used.
+
+* Add options "linear" and "linear_regional_average" for the `replace_NAs` argument.
+
+* Add vignettes
+
+* Remove dependency on the stringr package
+
+* Improve tests
+
+# GDPuc 0.5.1
 
 First public release.